+++ Samples/DemoScenes/Scripts/BehaviourTree/StandaloneBT/BTStandaloneBase.cs
@@ -1,14 +1,10 @@
 using BehaviourTree;
-using CommonCore;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace DemoScenes
 {
     public abstract class BTStandaloneBase : BTBrainBase
     {
-        [SerializeField] float DefaultInteractableSearchRange = 30f;
-
         protected override void OnPreTickBrain(float InDeltaTime)
         {
             base.OnPreTickBrain(InDeltaTime);
@@ -37,93 +33,5 @@ namespace DemoScenes
                     LinkedBlackboard.Set(CommonCore.Names.Target_GameObject, (GameObject)null);
             }
         }
-
-        #region Interactable Helpers
-        public float GetUseInteractableDesire(GameObject InQuerier, float InSearchRange)
-        {
-            // if we already have an interaction block switching
-            SmartObject TargetSO = null;
-            LinkedBlackboard.TryGet(CommonCore.Names.Interaction_SmartObject, out TargetSO, null);
-            if (TargetSO != null)
-            {
-                BaseInteraction TargetInteraction = null;
-                LinkedBlackboard.TryGet(CommonCore.Names.Interaction_Type, out TargetInteraction, null);
-                if (TargetInteraction != null)
-                {
-                    return 0.25f;
-                }
-            }
-
-            int NumCandidateInteractables = 0;
-
-            float MaxInteractionRange = InSearchRange > 0 ? InSearchRange : DefaultInteractableSearchRange;
-            float MaxInteractionRangeSq = MaxInteractionRange * MaxInteractionRange;
-
-            // loop through all of the smart objects
-            foreach (var CandidateSO in SmartObjectManager.Instance.RegisteredObjects)
-            {
-                // loop through all of the interactions
-                foreach (var CandidateInteraction in CandidateSO.Interactions)
-                {
-                    if (!CandidateInteraction.CanPerform())
-                        continue;
-
-                    if ((CandidateSO.InteractionPoint - transform.position).sqrMagnitude > MaxInteractionRangeSq)
-                        continue;
-
-                    if ((Navigation != null) && !Navigation.IsLocationReachable(transform.position, CandidateSO.InteractionPoint))
-                        continue;
-
-                    ++NumCandidateInteractables;
-                }
-            }
-
-            // no interactions?
-            if (NumCandidateInteractables == 0)
-            {
-                return float.MinValue;
-            }
-
-            return 0.25f;
-        }
-
-        public void SelectRandomInteraction(GameObject InQuerier, float InSearchRange, System.Action<SmartObject, BaseInteraction> InCallbackFn)
-        {
-            List<System.Tuple<SmartObject, BaseInteraction>> CandidateInteractions = new();
-
-            float MaxInteractionRange = InSearchRange > 0 ? InSearchRange : DefaultInteractableSearchRange;
-            float MaxInteractionRangeSq = MaxInteractionRange * MaxInteractionRange;
-
-            // loop through all of the smart objects
-            foreach (var CandidateSO in SmartObjectManager.Instance.RegisteredObjects)
-            {
-                // loop through all of the interactions
-                foreach (var CandidateInteraction in CandidateSO.Interactions)
-                {
-                    if (!CandidateInteraction.CanPerform())
-                        continue;
-
-                    if ((CandidateSO.InteractionPoint - transform.position).sqrMagnitude > MaxInteractionRangeSq)
-                        continue;
-
-                    if ((Navigation != null) && !Navigation.IsLocationReachable(transform.position, CandidateSO.InteractionPoint))
-                        continue;
-
-                    CandidateInteractions.Add(new System.Tuple<SmartObject, BaseInteraction>(CandidateSO, CandidateInteraction));
-                }
-            }
-
-            // no interactions?
-            if (CandidateInteractions.Count == 0)
-            {
-                InCallbackFn(null, null);
-                return;
-            }
-
-            // pick a random interaction
-            var SelectedIndex = CandidateInteractions.Count == 1 ? 0 : Random.Range(0, CandidateInteractions.Count);
-            InCallbackFn(CandidateInteractions[SelectedIndex].Item1, CandidateInteractions[SelectedIndex].Item2);
-        }
-        #endregion
     }
 }